revert unwanted changes and git ignore the version.txt file
