+++ .gitignore
@@ -98,6 +98,7 @@ app/src/main/assets/version.txt
 app/src/main/assets/dependencies_version.json
 /intellij.gdsl
 
+app/version.txt
 # Editor temporary files
 *~
 \#*#
